Create macro to generate AstClass enums
